Correct MSVC warning in clutils/dirobj.cc
Change int to size_t to avoid warning C4267: 'initializing' : conversion from 'size_t' to 'int', possible loss of data

+++ src/clutils/dirobj.cc
@@ -260,7 +260,7 @@ const char * DirObj::ValidDirectories( const char * path ) {
     static char buf[MAXPATHLEN + 1];
 #endif
     strcpy( buf, path );
-    int i = strlen( path );
+    size_t i = strlen( path );
 
     while( !IsADirectory( RealPath( buf ) ) && i >= 0 ) {
         for( --i; buf[i] != '/' && i >= 0; --i ) {
